requirements.txt
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.1
PyNaCl==1.5.0
base58==2.1.1
